bsserver-demo: move some byte utils function to byteutils.js
